Private/dkr defintions (#19)
* feat: initial implementation of the definitions feature

This commit contains an initial version of the definitions feature. It shows all definitions within ExpansionTiles. Some layout adjustments and documentation have to be done.

* feat: initial implementation of the definitions feature

This commit contains an initial version of the definitions feature. It shows all definitions within ExpansionTiles. Some layout adjustments and documentation have to be done.

* refactor: move content loader to specific feature

The content loader moved to the specific features in order to increase the performance when starting the app, caused by to huge json files, which have to be loaded from remote.
Added a missing SectionBuilder to the ExpansionTile1 widget. Did "dart format .".

* refactor: resovled merge issues

---------

Co-authored-by: Walnut IT <daniel.krentzlin@walnut-it.com>