feat: updated AppBarTemplate
The AppBarTemplate can differ now between large, medium and small screen size and adjusts its behaviour to it. Addtionally an about us button was added to the AppBarTemplate. If the screen size is small the about us button will disappear in a menu and when clicking the menu button then the drawer will open.